Add params to FunctionSet 53243
Motivation
------------
Patch 18.8.50 introduced a new item which uses FunctionSet 53243 which was previously only used by the "Charge!" nano.  Upon analyzing the data structure of the records for both items it becomes apparent that this function actually has a single 4-byte integer parameter rather than 0 parameters.  For Charge! this did not matter to the parser because it was the last function in the list of functions and the next 4 bytes were incidentally read and ignored, but for the new MA attack item from the high level totw instance that was added in 18.8.50 this function appears halfway through the list and causes a parsing error.

Modifications
---------------
Updated the FunctionSets config for 53243 to specify a single 4-byte integer parameter.